+++ CHANGELOG.md
@@ -37,5 +37,5 @@
 * Add New feature for Animation.
 * Add New feature for Dependency injection.
 * Add New feature for Utils.
-* Add New feature for State management.
+* Add New feature for State management.     
 * Fix Bug GetBuilder.